chore: renamed List to Column
